feat: Add support for programmatic definitions via .zoo.yml file
* feat(services): .zoo.yml support

Implements a scheduled task that reads .zoo.yml file from a repository root and updates the zoo registry.
This file contains data about the repository. This is the same data that you can manually add through zoo
application form.

Description of the .zoo.yml format can be found in the README.md file

Co-authored-by: Pavel Dedik <dedikx@gmail.com>